Modified field read function + small fixes + simplified/removed some tests in order to improve testing times
